requirements.txt
python-nmap
requests
jinja2